fixed chris's errors. ;)

+++ source/kernel/gdt.c
@@ -50,7 +50,7 @@ void gdt_install()
 {
     /* Setup the GDT pointer and limit */
     gp.limit = (sizeof(struct gdt_entry) * 3) - 1;
-    gp.base = &gdt;
+    gp.base = (UINT)&gdt;
 
     //! null descriptor
     gdt_set_gate(0, 0, 0, 0, 0);
